pyproject.toml
[project]
name = "rag-genai-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.15",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asgiref==3.8.1",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "beautifulsoup4==4.13.3",
    "build==1.2.2.post1",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "chroma-hnswlib==0.7.6",
    "chromadb==0.6.3",
    "click==8.1.8",
    "coloredlogs==15.0.1",
    "dataclasses-json==0.6.7",
    "deprecated==1.2.18",
    "distro==1.9.0",
    "docstring-parser==0.16",
    "durationpy==0.9",
    "esprima==4.0.1",
    "fastapi==0.115.12",
    "filelock==3.18.0",
    "flatbuffers==25.2.10",
    "frozenlist==1.5.0",
    "fsspec==2025.3.2",
    "google-api-core==2.24.2",
    "google-auth==2.38.0",
    "google-cloud-aiplatform==1.87.0",
    "google-cloud-bigquery==3.31.0",
    "google-cloud-core==2.4.3",
    "google-cloud-resource-manager==1.14.2",
    "google-cloud-storage==2.19.0",
    "google-crc32c==1.7.1",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.69.2",
    "grpc-google-iam-v1==0.14.2",
    "grpcio==1.71.0",
    "grpcio-status==1.71.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.30.1",
    "humanfriendly==10.0",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "importlib-resources==6.5.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "kubernetes==32.0.1",
    "langchain==0.3.22",
    "langchain-chroma==0.2.2",
    "langchain-community==0.3.20",
    "langchain-core==0.3.49",
    "langchain-google-vertexai==2.0.18",
    "langchain-text-splitters==0.3.7",
    "langgraph==0.3.21",
    "langgraph-checkpoint==2.0.23",
    "langgraph-prebuilt==0.1.7",
    "langgraph-sdk==0.1.60",
    "langsmith==0.3.21",
    "markdown-it-py==3.0.0",
    "marshmallow==3.26.1",
    "mdurl==0.1.2",
    "mmh3==5.1.0",
    "monotonic==1.6",
    "mpmath==1.3.0",
    "multidict==6.3.0",
    "mypy-extensions==1.0.0",
    "numpy==1.26.4",
    "oauthlib==3.2.2",
    "onnxruntime==1.21.0",
    "opentelemetry-api==1.31.1",
    "opentelemetry-exporter-otlp-proto-common==1.31.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.31.1",
    "opentelemetry-instrumentation==0.52b1",
    "opentelemetry-instrumentation-asgi==0.52b1",
    "opentelemetry-instrumentation-fastapi==0.52b1",
    "opentelemetry-proto==1.31.1",
    "opentelemetry-sdk==1.31.1",
    "opentelemetry-semantic-conventions==0.52b1",
    "opentelemetry-util-http==0.52b1",
    "orjson==3.10.16",
    "ormsgpack==1.9.1",
    "overrides==7.7.0",
    "packaging==24.2",
    "posthog==3.23.0",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.4",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.1",
    "pydantic-core==2.33.0",
    "pydantic-settings==2.8.1",
    "pygments==2.19.1",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.0.0",
    "rsa==4.9",
    "shapely==2.0.7",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sqlalchemy==2.0.40",
    "starlette==0.46.1",
    "sympy==1.13.3",
    "tenacity==9.0.0",
    "tokenizers==0.21.1",
    "tqdm==4.67.1",
    "typer==0.15.2",
    "typing-extensions==4.13.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.0",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "validators==0.34.0",
    "watchfiles==1.0.4",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "xxhash==3.5.0",
    "yarl==1.18.3",
    "zipp==3.21.0",
    "zstandard==0.23.0",
]
